Attempting fix for travis testing exit code issue

+++ build_tests.sh
@@ -9,10 +9,18 @@ cd ../vendor
 node api.js &
 export VENDOR_PID=$!
 cd ../backend
+node server.js &
+export BACKEND_PID=$!
+
+cd ../backend
 npm run test
+if [ $? -ne 0] 
+then
+    exit $?
 
 cd ../frontend
 npm test -- --watchAll=false
 npm run cy:run
 
 kill $VENDOR_PID
+kill $BACKEND_PID
